pyproject.toml
[project]
name = "fcp-server"
version = "1.0.0"
description = "FCP Reference Server - Food Context Protocol implementation with Gemini 3"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Food Context Protocol Contributors" }
]
keywords = ["food", "ai", "gemini", "mcp", "protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core AI
    "google-genai>=1.0.0",
    "pydantic-ai>=0.1.0",
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.22",
    "slowapi>=0.1.9",
    # MCP Protocol
    "mcp>=1.0.0",
    # Data Validation
    "pydantic>=2.0.0",
    # HTTP Client
    "httpx>=0.27.0",
    # CLI
    "typer>=0.12.0",
    "rich>=13.0.0",
    # Utilities
    "tenacity>=8.0.0",
    "anyio>=4.0.0",
    "limits>=3.7.0",
    "typing-extensions>=4.5.0",
    "apscheduler>=3.10.0",
    "sniffio>=1.3.0",
    "distro>=1.9.0",
    "deprecated>=1.2.0",
    "packaging>=23.0",
    "wrapt>=1.14.0",
    # Local persistence
    "aiosqlite>=0.20.0",
    # Prometheus metrics (lightweight counters only)
    "prometheus-client>=0.20.0",
    # Logfire for enhanced observability
    "logfire>=1.0.0",
    "pydantic-settings>=2.12.0",
    "pillow>=12.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "pytest-timeout>=2.4.0",
    "freezegun>=1.5.5",
    "respx>=0.22.0",
    "ruff>=0.1.0",
    "watchfiles>=1.0.0",
]
firestore = [
    "google-cloud-firestore>=2.14.0",
    "google-cloud-storage>=2.14.0",
]

[project.scripts]
fcp-server = "fcp.server:main"

[project.urls]
Homepage = "https://github.com/Food-Context-Protocol/fcp-gemini-server"
Documentation = "https://docs.fcp.dev"
Repository = "https://github.com/Food-Context-Protocol/fcp-gemini-server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fcp"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = ["src/fcp/client"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long - handled by formatter

[tool.ty]
# ty type checker configuration

[tool.ty.rules]
# Ignore unresolved-attribute for dynamic google.genai types
unresolved-attribute = "ignore"
# Ignore invalid-argument-type for flexible Pydantic patterns
invalid-argument-type = "ignore"
# Ignore unused-ignore-comment
unused-ignore-comment = "ignore"
# Ignore unresolved-import for conditional/lazy exports
unresolved-import = "ignore"
# Ignore missing-argument for optional kwargs
missing-argument = "ignore"
# Ignore unknown-argument for generated client mismatches
unknown-argument = "ignore"
# Ignore possibly-missing-attribute for optional types
possibly-missing-attribute = "ignore"
# Ignore not-iterable for dynamic response parsing
not-iterable = "ignore"

[tool.ty.src]
exclude = ["src/fcp/client"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-v --tb=short --asyncio-mode=auto"
timeout = 10
timeout_method = "thread"
markers = [
    "asyncio: async test marker",
    "small: fast, hermetic unit tests",
    "medium: tests that may touch filesystem or heavier fixtures",
    "large: slow tests (integration/e2e)",
    "integration: tests that require external services or emulators",
    "e2e: end-to-end tests",
    "slow: explicitly slow tests",
    "property: property-based tests",
    "sdk: tests that exercise the SDK client surface",
    "network: tests that require network access",
    "gemini: tests that exercise Gemini-backed features",
    "firestore: tests that require Firestore/database access",
]
filterwarnings = [
    "error",
    "ignore:Inheritance class AiohttpClientSession from ClientSession is discouraged:DeprecationWarning",
    "ignore:Interactions usage is experimental and may change in future versions\\.:UserWarning",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "tests/*",
    ".venv/*",
    "src/fcp/client/*",
    "src/fcp/scheduler/jobs.py",
    "src/fcp/services/gemini_generation.py",
    "src/fcp/services/gemini_helpers.py",
    "src/fcp/mcp/protocols.py",  # Protocol interfaces only, no executable code
    "src/fcp/services/cloud_storage_backend.py",  # Production-only GCS code requiring google-cloud-storage
    "src/fcp/services/firestore_backend.py",  # Pre-existing 90 test failures - needs fixture refactor
]

[tool.coverage.report]
fail_under = 100
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[dependency-groups]
dev = [
    "freezegun>=1.5.5",
    "hypothesis>=6.151.5",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-timeout>=2.4.0",
    "respx>=0.22.0",
    "ruff>=0.15.0",
    "sourcery>=1.43.0",
    "ty>=0.0.14",
]
